requirements.txt

streamlit
pandas
google-generativeai
python-dotenv
Pillow
numpy
chromadb
scikit-learn
altair
matplotlib
wordcloud
tiktoken